Refactor BrowserDriverManager

+++ actions/commons/BaseTestMultipleBrowserEnvironment.java
@@ -14,7 +14,7 @@ import org.testng.Reporter;
 
 import factoryEnvironment.BrowserstackFactory;
 import factoryEnvironment.CrossbrowserFactory;
-import factoryEnvironment.ENVIRONMENT;
+import factoryEnvironment.EnvironmentList;
 import factoryEnvironment.GridFactory;
 import factoryEnvironment.LambdaFactory;
 import factoryEnvironment.LocalFactory;
@@ -60,7 +60,7 @@ public class BaseTestMultipleBrowserEnvironment {
 
 	protected String getEnvironmentUrl(String serverName) {
 		String envUrl = null;
-		ENVIRONMENT environment = ENVIRONMENT.valueOf(serverName.toUpperCase());
+		EnvironmentList environment = EnvironmentList.valueOf(serverName.toUpperCase());
 
 		switch (environment) {
 		case DEV:
@@ -89,51 +89,53 @@ public class BaseTestMultipleBrowserEnvironment {
 		return this.driver;
 	}
 
-	protected void closeBrowserDriver() {
-		String cmd = null;
-		try {
-			String osName = System.getProperty("os.name").toLowerCase();
-			log.info("OS name = " + osName);
-
-			String driverInstanceName = driver.toString().toLowerCase();
-			log.info("Driver instance name = " + driverInstanceName);
-
-			String browserDriverName = null;
-
-			if (driverInstanceName.contains("chrome")) {
-				browserDriverName = "chromedriver";
-			} else if (driverInstanceName.contains("internetexplorer")) {
-				browserDriverName = "IEDriverServer";
-			} else if (driverInstanceName.contains("firefox")) {
-				browserDriverName = "geckodriver";
-			} else if (driverInstanceName.contains("edge")) {
-				browserDriverName = "msedgedriver";
-			} else if (driverInstanceName.contains("opera")) {
-				browserDriverName = "operadriver";
-			} else {
-				browserDriverName = "safaridriver";
-			}
-
-			if (osName.contains("window")) {
-				cmd = "taskkill /F /FI \"IMAGENAME eq " + browserDriverName + "*\"";
-			} else {
-				cmd = "pkill " + browserDriverName;
-			}
-
-			if (driver != null) {
-				driver.manage().deleteAllCookies();
-				driver.quit();
-			}
-		} catch (Exception e) {
-			log.info(e.getMessage());
-		} finally {
+	protected void closeBrowserDriver(String envName) {
+		if (envName.contains("local") || envName.contains("grid")) {
+			String cmd = null;
 			try {
-				Process process = Runtime.getRuntime().exec(cmd);
-				process.waitFor();
-			} catch (IOException e) {
-				e.printStackTrace();
-			} catch (InterruptedException e) {
-				e.printStackTrace();
+				String osName = System.getProperty("os.name").toLowerCase();
+				log.info("OS name = " + osName);
+
+				String driverInstanceName = driver.toString().toLowerCase();
+				log.info("Driver instance name = " + driverInstanceName);
+
+				String browserDriverName = null;
+
+				if (driverInstanceName.contains("chrome")) {
+					browserDriverName = "chromedriver";
+				} else if (driverInstanceName.contains("internetexplorer")) {
+					browserDriverName = "IEDriverServer";
+				} else if (driverInstanceName.contains("firefox")) {
+					browserDriverName = "geckodriver";
+				} else if (driverInstanceName.contains("edge")) {
+					browserDriverName = "msedgedriver";
+				} else if (driverInstanceName.contains("opera")) {
+					browserDriverName = "operadriver";
+				} else {
+					browserDriverName = "safaridriver";
+				}
+
+				if (osName.contains("window")) {
+					cmd = "taskkill /F /FI \"IMAGENAME eq " + browserDriverName + "*\"";
+				} else {
+					cmd = "pkill " + browserDriverName;
+				}
+
+				if (driver != null) {
+					driver.manage().deleteAllCookies();
+					driver.quit();
+				}
+			} catch (Exception e) {
+				log.info(e.getMessage());
+			} finally {
+				try {
+					Process process = Runtime.getRuntime().exec(cmd);
+					process.waitFor();
+				} catch (IOException e) {
+					e.printStackTrace();
+				} catch (InterruptedException e) {
+					e.printStackTrace();
+				}
 			}
 		}
 	}
